fix: add correct structure
