+++ src/components/Main.tsx
@@ -1,10 +1,12 @@
 import { useState, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import Search from './Search/Search';
-import Result from './Result/Result';
+import PlanetList from './PlanetList/PlanetList';
 import Universe from './Universe/Universe';
 import './Main.css';
 import Pagination, { PageInfo } from './Pagination/Pagination';
+import { fetchDataForPlanetList } from '../utils/dataFetcher';
+import { PlanetInfo } from './PlanetList/PlanetInfoTypes';
 
 type Props = {
   children?: JSX.Element;
@@ -12,7 +14,9 @@ type Props = {
 
 const Main: React.FC<Props> = () => {
   const [searchQuery, setSearchQuery] = useState<string>('empty');
-
+  const [error, setError] = useState<Error | null>(null);
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [items, setItems] = useState<PlanetInfo[]>([]);
   const [pageInfo] = useState<PageInfo>({
     count: 0,
     next: null,
@@ -25,9 +29,16 @@ const Main: React.FC<Props> = () => {
 
   const updateSearchQuery = (value: string) => {
     setSearchQuery(value);
+    fetchDataForList(value);
   };
 
-  useEffect(() => {}, [currentPage]);
+  const fetchDataForList = (query: string) => {
+    fetchDataForPlanetList(query, setIsLoaded, setItems, setError);
+  };
+
+  useEffect(() => {
+    fetchDataForList(searchQuery);
+  }, [currentPage, searchQuery]);
 
   const handleNavigate = (newPage: number) => {
     setSearchParams({ page: newPage.toString() });
@@ -41,7 +52,12 @@ const Main: React.FC<Props> = () => {
       <div className="search-result">
         {' '}
         <Search onSearchSubmit={updateSearchQuery} />
-        <Result data={searchQuery} />
+        <PlanetList
+          data={searchQuery}
+          isLoaded={isLoaded}
+          items={items}
+          error={error}
+        />
         <Pagination
           currentPage={currentPage}
           info={pageInfo}
