+++ src/components/PlanetList/NotFoundMessage.tsx
@@ -1,0 +1,5 @@
+const NotFoundMessage: React.FC = () => {
+  return <p className="search__text">Nothing found</p>;
+};
+
+export default NotFoundMessage;
