+++ src/components/PlanetList/PlanetInfoTypes.tsx
@@ -1,0 +1,11 @@
+export interface PlanetInfo {
+  name: string;
+  rotation_period: string;
+  orbital_period: string;
+  diameter: string;
+  climate: string;
+  gravity: string;
+  population: string;
+  terrain: string;
+  surface_water: string;
+}
