+++ src/components/PlanetList/PlanetList.css
@@ -19,6 +19,9 @@
   width: 80%;
   color: white;
   font-family: 'Courier New', Courier, monospace;
+  border: 2px solid #603f8b;
+  border-radius: 4px;
+  padding: 20px;
 }
 
 .result__wrapper--all {
